Move class into cartesian.py
